+++ src/main/java/dev/willyelton/crystal_tools/common/events/BlockEvents.java
@@ -29,6 +29,7 @@ import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.level.BlockDropsEvent;
 import net.neoforged.neoforge.event.level.BlockEvent;
+import net.neoforged.neoforge.event.level.block.BreakBlockEvent;
 
 import java.util.List;
 import java.util.Optional;
@@ -36,7 +37,7 @@ import java.util.Optional;
 @EventBusSubscriber(modid = CrystalTools.MODID)
 public class BlockEvents {
     @SubscribeEvent
-    public static void breakEvent(BlockEvent.BreakEvent event) {
+    public static void breakEvent(BreakBlockEvent event) {
         Player player = event.getPlayer();
 
         // Backpack levels